Fixes the bug blocking Works from Collection Page (#1680).
